Updated Source + Jar

+++ Source Code/Sparkz Enginev0.08 - 0.8/src/com/engine/ParticleTypes/Emitter.java
@@ -8,11 +8,11 @@ public class Emitter extends Molecule {
     public int life = 800;
 
     public Emitter() {super();}
-    public Emitter(double _x, double _y, double _radius, double speed, int direction) {
+    public Emitter(double _x, double _y, double _radius, double speed, double direction) {
         super(_x, _y, _radius, speed, direction, 0);
     }
 
-    private void emitAtPoint() {ParticleModes.fireworksMode(x, y, fireworksWind, (int) (fireworksSpeedVal), fireworksAmount);}
+    private void emitAtPoint() {ParticleModes.fireworksMode(x, y, fireworksWind, fireworksSpeedVal, fireworksAmount);}
     public void render() {
         graphics2D.setColor(color);
     }
